genericize fetchTileData
