+++ src/fetchTileData.ts
@@ -2,11 +2,12 @@
 
 import fnv1a from '@sindresorhus/fnv1a';
 import {LRUCache} from 'lru-cache';
-import {s3TileFetcher} from './s3TileFetcher.js';
 
-export const TILE_MISSING = Symbol();
+type Fetcher<T extends unknown[]> = (...args: T) => Promise<Buffer>;
+
+export const TILE_MISSING = Symbol(); // MUST be thrown by `Fetcher` to indicate a missing tile
 export const BAD_TILE = Symbol();
-// ... other error conditions, e.g. permissions, network failure, etc. ...
+// ... could signal other error conditions, e.g. bad permissions, network failure, etc.
 
 const cache = new LRUCache<bigint, Buffer>({
     maxSize: 500_000_000, // 500 MB
@@ -19,73 +20,33 @@ const pending = new Map<bigint, Promise<Buffer>>();
 
 const missing = new Set<bigint>();
 
-export async function fetchTileData(Bucket: string, Key: string): Promise<Buffer> {
-    const hash = fnv1a(JSON.stringify([Bucket, Key]));
-
-    if (cache.has(hash)) return cache.get(hash)!;
-    if (pending.has(hash)) return pending.get(hash)!;
-
-    if (missing.has(hash)) throw TILE_MISSING;
-
-    const start = process.hrtime.bigint();
-
-    pending.set(
-        hash,
-        s3TileFetcher(Bucket, Key)
-            .then((buffer) => {
-                const ms = Number((process.hrtime.bigint() - start) / 1_000_000n);
-                if (ms > 1000) {
-                    console.log(`Loading s3://${Bucket}/${Key} took ${(ms / 1_000).toFixed(3)}s`);
-                }
-
-                // ...so future requests can use the cache
-                cache.set(hash, buffer);
-
-                // ...so pending requests will get the resolved value
-                return buffer;
-            })
-            .catch((error: unknown) => {
-                if (error === TILE_MISSING) missing.add(hash);
-                throw error;
-            })
-            .finally(() => {
-                // must clean up, else we will leak memory, and make the LRU `maxSize` pointless
-                pending.delete(hash);
-            }),
-    );
-
-    return pending.get(hash)!;
-}
-
-/*
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-type Fetcher<T extends any[]> = (...args: T) => Promise<Buffer>;
-
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-export async function fetchTileDataGeneric<T extends any[]>(fetcher: Fetcher<T>, args: T) {
+export async function fetchTileData<T extends unknown[]>(fetcher: Fetcher<T>, ...args: T) {
     const hash = fnv1a(JSON.stringify(args));
 
+    // If the LRU cache has it, return it
     if (cache.has(hash)) return cache.get(hash)!;
+
+    // If we're currently fetching it, return the promise
     if (pending.has(hash)) return pending.get(hash)!;
 
+    // If we know we can't get it, don't bother trying again
     if (missing.has(hash)) throw TILE_MISSING;
 
     pending.set(
         hash,
         fetcher(...args)
             .then((buffer) => {
-                cache.set(hash, buffer);
-                return buffer;
+                cache.set(hash, buffer); // ...so future requests can use the cache
+                return buffer; // ...so pending requests will get the resolved value
             })
             .catch((error: unknown) => {
                 if (error === TILE_MISSING) missing.add(hash);
                 throw error;
             })
             .finally(() => {
-                pending.delete(hash);
+                pending.delete(hash); // ... else we will leak memory & LRU cache is pointless
             }),
     );
 
     return pending.get(hash)!;
 }
-*/
